feat: Try upgrading a prefixes & Start testing compound units updating. So far we: - Check if the factor for the unit matches a prefix, if yes, assign it - Check if we can find a lower prefix to assign - If there are factors unaccounted for, try upgrading prefixes to pick up the factor
Next step is to upgrade a prefix while downgrading another. For example if unit A has a 1e5 exponent and unit B a 1e3 exponent, unit A needs to move to a 1e6 exponent and unit B to 1e2 so unit A gets the prefix Mega and unit B the prefic hecto.
Then if that's not enough, we need to try splitting the units that have an exponent. for example 1e5 A^2 = 1e3 A * 1e2 A = kA * hA
Then if that's not enough, that means there units canceling out, so we need to introduce a ratio of the unit that self cancels out as a last resort, just to pick up the prefactor.